backend/requirements.txt
asgiref==3.8.1
Django==5.0.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.0
pillow==10.3.0
pycryptodome==3.20.0
